+++ Generativeaiagent/models/UpdateDataSourceDetails.cs
@@ -38,6 +38,12 @@ namespace Oci.GenerativeaiagentService.Models
         public DataSourceConfig DataSourceConfig { get; set; }
         
         /// <value>
+        /// Key-value pairs to allow additional configurations.
+        /// </value>
+        [JsonProperty(PropertyName = "metadata")]
+        public System.Collections.Generic.Dictionary<string, string> Metadata { get; set; }
+        
+        /// <value>
         /// Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace.
         /// For more information, see [Resource Tags](https://docs.cloud.oracle.com/iaas/Content/General/Concepts/resourcetags.htm).
         /// <br/>