+++ Stackmonitoring/models/CreateMonitoredResourceTypeDetails.cs
@@ -49,7 +49,7 @@ namespace Oci.StackmonitoringService.Models
         public string Description { get; set; }
         
         /// <value>
-        /// The [OCID](https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm) of the tenancy containing the resource type.
+        /// Compartment Identifier [OCID](https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm).
         /// 
         /// </value>
         /// <remarks>